+++ package.json
@@ -218,6 +218,7 @@
     "@types/leaflet": "^1.7.5",
     "@types/leaflet-draw": "^0.4.14",
     "@types/leaflet.heat": "^0.2.1",
+    "@types/leaflet.markercluster": "^1.5.4",
     "@types/lodash.clonedeep": "^4.5.9",
     "@types/lodash.debounce": "^4.0.9",
     "@types/lodash.orderby": "^4.6.9",
